Commented out mean_diel -- Don't pull!
This one shouldn’t be merged with the master branch!
I haven’t dug into this, but it was trying to evaluate qmean_can_prss
which didn’t exist and would cause a segfault crash even though I don’t
think I was writing the diurnal cycle, so I just commented this out.