generate_deps.sh update for Mac + NetCDF libraries update
The “sed” command in generate_deps.sh fails on Unix, it expects an
extension as an argument. Also changed wording so it does not remove
netcdf.mod from mod_netcdf.mod in dependency.mk.
Changes in the include.mk template are for if you’re using the latest
NetCDF install (separate C and Fortran installations, I have NetCDF
4.3.3.1 for C and NetCDF 4.4.2 for Fortran)