Several changes to the decomposition scheme.  Most of them are minor. 1.  DECOMP_SCHEME = 2 now features a microbial soil carbon, which is closer to CENTURY's classification.  This also separates above- and below-ground fast (metabolic) and structural (woody debris) so fires can burn the above-ground component only.  The model now has 6 soil C pools: fast (ground and below-ground), structural (ground and below-ground), microbial and slow.  Nitrogen fast and structural pools were also expanded into ground and below-ground, and now the code has an explicit structural N pool.
2.  To keep damage (such as liana 'pruning') more tractable, bark and heartwood were
split into above- and below-ground everywhere in the code.

3.  To help understanding the contributions of heterotrophic respiration, the contribution
from each of the 6 pools are now part of the output

4.  Litterfall was standardised across the model.  Some routines used to dump all leaves
into fast, whilst others would separate labile from lignified.  Now every routine sorts
litter in the same way.  A similar ratio fraction of labile was defined for woody tissues,
and implemented throughout the code although it is currently set to zero.