Initialize biomass pools before other cohort vars
In reproduction_driver code. This is because `plant_hydro::rwc2tw`
requires biomass variables for scaling, but `cpatch` gets wiped and is
not initialized in time, resulting in a floating point arithmetic
error (biomass values are allocated but uninitialized, so they have
nonsense values like 1e-44).

This just moves the initialization of cpatch biomass values from
recruits before the initialization of other variables.