Sanitize history.xml string output
`len(value)` > `len(trim(value))`, so when we tell Fortran to write a
string of length `len(value)` but only give it `len(trim(value))`
characters, it fills the rest with unreadalbe garbage that trips up most
file readers. This small fix ensures that we only write out as many
characters as we mean to.