Adding missing variable FC_TYPE to include.mk.docker, and temporarily disabling MPI settings in the same file.
MPI is giving some odd messages in the CI (like failing in fatal_error.F90 when trying
to compile call to MPI_Abort). The call to MPI_Abort seems correct, so this may be a problem
in the MPI side, not ED2. For now we disable it as very few people are using truly regional
runs nowadays, but if anyone knows a solution, please feel free to revise the code and
revert the CI and Docker CI compilations to allow MPI.