Minor bug fix to prevent CI tests to fail, and update to dataset sources for the test suite.
Somehow the CI tests were crashing due to a "merge" command in the inventory initialisation.
Switching the merge with an equivalent "where" construction seems to fix it. This could be
a compiler issue, but if the problem pops up again, we may want to deprecate merge in the
code and go with where instead.